Refactor to use strings for prices

+++ src/backend/src/main/java/com/bill_split/app/data/Item.java
@@ -24,7 +24,7 @@ public class Item {
   public Item() {
   }
 
-  public Item(String name, Long cost) {
+  public Item(String name, BigDecimal cost) {
     this.name = name;
     this.cost = cost;
   }
